Move public router instance to module level
There's only one instance for the whole app, and it conceptually lives
outside of the React tree. We don't need to create it during render.
