+++ packages/next/src/client/components/app-router-instance.ts
@@ -6,10 +6,24 @@ import {
   ACTION_SERVER_ACTION,
   ACTION_NAVIGATE,
   ACTION_RESTORE,
+  type NavigateAction,
+  ACTION_HMR_REFRESH,
+  PrefetchKind,
+  ACTION_PREFETCH,
 } from './router-reducer/router-reducer-types'
 import { reducer } from './router-reducer/router-reducer'
 import { startTransition } from 'react'
 import { isThenable } from '../../shared/lib/is-thenable'
+import { prefetch as prefetchWithSegmentCache } from './segment-cache'
+import { dispatchAppRouterAction } from './use-action-queue'
+import { addBasePath } from '../add-base-path'
+import { createPrefetchURL, isExternalURL } from './app-router'
+import { prefetchReducer } from './router-reducer/reducers/prefetch-reducer'
+import type {
+  AppRouterInstance,
+  NavigateOptions,
+  PrefetchOptions,
+} from '../../shared/lib/app-router-context.shared-runtime'
 
 export type DispatchStatePromise = React.Dispatch<ReducerState>
 
@@ -210,4 +224,113 @@ export function createMutableActionQueue(
 
 export function getCurrentAppRouterState(): AppRouterState | null {
   return globalActionQueue !== null ? globalActionQueue.state : null
+}
+
+function getAppRouterActionQueue(): AppRouterActionQueue {
+  if (globalActionQueue === null) {
+    throw new Error(
+      'Internal Next.js error: Router action dispatched before initialization.'
+    )
+  }
+  return globalActionQueue
+}
+
+function dispatchNavigateAction(
+  href: string,
+  navigateType: NavigateAction['navigateType'],
+  shouldScroll: boolean
+): void {
+  // TODO: This stuff could just go into the reducer. Leaving as-is for now
+  // since we're about to rewrite all the router reducer stuff anyway.
+  const url = new URL(addBasePath(href), location.href)
+  if (process.env.__NEXT_APP_NAV_FAIL_HANDLING) {
+    window.next.__pendingUrl = url
+  }
+  dispatchAppRouterAction({
+    type: ACTION_NAVIGATE,
+    url,
+    isExternalUrl: isExternalURL(url),
+    locationSearch: location.search,
+    shouldScroll,
+    navigateType,
+    allowAliasing: true,
+  })
+}
+
+/**
+ * The app router that is exposed through `useRouter`. These are public API
+ * methods. Internal Next.js code should call the lower level methods directly
+ * (although there's lots of existing code that doesn't do that).
+ */
+export const publicAppRouterInstance: AppRouterInstance = {
+  back: () => window.history.back(),
+  forward: () => window.history.forward(),
+  prefetch: process.env.__NEXT_CLIENT_SEGMENT_CACHE
+    ? // Unlike the old implementation, the Segment Cache doesn't store its
+      // data in the router reducer state; it writes into a global mutable
+      // cache. So we don't need to dispatch an action.
+      (href: string, options?: PrefetchOptions) => {
+        const actionQueue = getAppRouterActionQueue()
+        prefetchWithSegmentCache(
+          href,
+          actionQueue.state.nextUrl,
+          actionQueue.state.tree,
+          options?.kind === PrefetchKind.FULL
+        )
+      }
+    : (href: string, options?: PrefetchOptions) => {
+        // Use the old prefetch implementation.
+        const actionQueue = getAppRouterActionQueue()
+        const url = createPrefetchURL(href)
+        if (url !== null) {
+          // The prefetch reducer doesn't actually update any state or
+          // trigger a rerender. It just writes to a mutable cache. So we
+          // shouldn't bother calling setState/dispatch; we can just re-run
+          // the reducer directly using the current state.
+          // TODO: Refactor this away from a "reducer" so it's
+          // less confusing.
+          prefetchReducer(actionQueue.state, {
+            type: ACTION_PREFETCH,
+            url,
+            kind: options?.kind ?? PrefetchKind.FULL,
+          })
+        }
+      },
+  replace: (href: string, options?: NavigateOptions) => {
+    startTransition(() => {
+      dispatchNavigateAction(href, 'replace', options?.scroll ?? true)
+    })
+  },
+  push: (href: string, options?: NavigateOptions) => {
+    startTransition(() => {
+      dispatchNavigateAction(href, 'push', options?.scroll ?? true)
+    })
+  },
+  refresh: () => {
+    startTransition(() => {
+      dispatchAppRouterAction({
+        type: ACTION_REFRESH,
+        origin: window.location.origin,
+      })
+    })
+  },
+  hmrRefresh: () => {
+    if (process.env.NODE_ENV !== 'development') {
+      throw new Error(
+        'hmrRefresh can only be used in development mode. Please use refresh instead.'
+      )
+    } else {
+      startTransition(() => {
+        dispatchAppRouterAction({
+          type: ACTION_HMR_REFRESH,
+          origin: window.location.origin,
+        })
+      })
+    }
+  },
+}
+
+// Exists for debugging purposes. Don't use in application code.
+if (typeof window !== 'undefined' && window.next) {
+  window.next.router = publicAppRouterInstance
 }
